MMCSD SDIO:  Add support for a new SDWAIT_WRCOMPLETE condition.  The previous logic used a busy-wait loop to pool the card R1 start to determine when the card was ready for the next transfer.  That busy-wait can be quite long -- hundreds of milliseconds.  And alternative is to look the the SD D0 pin which will change state when the card is no longer busy.
This logic implements a change the avoids the busy-wait poll by reconfiguring the SD D0 pin as a GPIO interrupt, then waiting for the card to becom ready without taking up CPU cycles.

This change is conditioned on CONFIG_MMCSD_SDIOWATI_WRCOMPLETE and is currenlty only implemented for the STM32 SDIO driver.

From David Sidrane