Protected mode:  Redesign how the user space heap is accessed from the kernel code.  It used to call memory management functions in user space via function pointers in the userspace interface.  That is inefficient because the first thing that those memory management functions do is to trap back into the kernel to get the current PID.  Worse, that operation can be fatal is certain fragile situations such as when a task is exitting.
The solution is to remove all of the memory management function calls from the interface.  Instead, the interface exports the userspace heap structure and then kernel size implementations of those memory management functions will operate on the userspace heap structure.  This avoids the unnecessary system calls and, more importantly, failures do to freeing memory when a test exits.